Assignment 2v1.0
Program is running now and completing objectives of assignments. Fixed
issue with inventory in Vendor. Added functions in Vendor for
PrintInventory, SellItem, OpenShop, and CloseShop. Most functions are
private and are called from the one public function of OpenShop. Input
for choosing index of item and quantity are both separately validated
and will lead back to either the menu or choosing an item again. Vendor
also has a method for closing the shop which is told the player after
every transaction. After selecting an item and quantity, the total price
and how many were sold are shown to the player.